backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
python-jose
python-dotenv